requirements.txt
aixplain
python-dotenv
requests
beautifulsoup4
lxml
kaggle
tqdm
rich
discord.py
